Cannot make emergency call from decryption screen
Step to reproduce it on Nexus 4 with 4.4.2:
1. Encrypt the device in Setting
2. Device will reboot into decryption UI
3. Make an emergency call

After the phone is encrypted, bootup into decryption UI.
PackageManagerService just loads core App at this time. To
make an emergency call, Telephony service starts
CallHandlerService whice is defined in InCallUI. InCallUI
is marked as a core app in it's AndroidManifest.xml file,
but it's not used. The InCallUI's codes is included by
Dialer app which is not core app, so the service cannot be
launched, the emergency will fail.
To address the issue, just add Dialer app to the core apps.

Also include another null protection
Since contacts provider is not a core app, contact ID query
will return null cursor. So, validate the cursor before
using it.

Change-Id: I189b668079f3d40512d72767a41c2832d29aa111
Signed-off-by: wangruix <ruix.wang@intel.com>
Signed-off-by: Guobin Zhang <guobin.zhang@intel.com>