Fixing CallLog prematurely mark as read all voicemails
A voicemail should be marked as read after the user has really read the content.
Upon entering the CallLog, the Dialog marks whole lot voicemails prematurely
as read, regardless of the content availability, nor the played/not yet played status.
This patch marks as read the voicemail only after its content is prepared, and cleans
up the code that is now unused.

Change-Id: Ia51bee12d73fd5dc2a7333362d35494579ed2ec0
Signed-off-by: HUANG Menghuai <menghuai.huang@orange.com>